vere: removes urbit-worker
